_backup/transcritor-pdf/requirements.txt
# -*- coding: utf-8 -*-
# Dependências Python para transcritor-pdf

# === Core ===
python-dotenv   # Para carregar variáveis de ambiente do .env
tenacity            # Para lógica de retentativa (ex: em chamadas de API)

# === Langchain Core & LLM/Embedding Integration ===
langchain           # Framework principal
langchain-openai    # Integração com OpenAI (ChatModels, Embeddings)
langchain-community # Integrações da comunidade (ex: FakeListChatModel para testes)
openai              # Biblioteca cliente da OpenAI (usada por langchain-openai)
google-generativeai # Para acesso direto à API Gemini (opcional)

# === PDF & Image Handling ===
pypdfium2           # Manipulação de PDF (dividir páginas)
Pillow              # Manipulação de Imagem (base, usada por pypdfium2/skimage)
scikit-image        # Processamento avançado de imagem (CLAHE, Sauvola, Median, etc.)
docling             # For advanced PDF layout analysis and document understanding

# === Vector Store / Database ===
asyncpg             # Driver Async PostgreSQL (para pgvector)
# langchain-postgres # Integração Langchain específica para PGVector (considerar depois)

# === API & Web Server ===
fastapi             # Framework para construção de APIs
uvicorn[standard]   # Servidor ASGI para FastAPI (com extras como uvloop, httptools)

# === Testing ===
pytest              # Framework de teste
pytest-mock         # Fixture para mocking/patching em testes
pytest-asyncio      # Plugin para suportar testes async com pytest
httpx               # Cliente HTTP assíncrono, dependência do TestClient do FastAPI
requests            # For simple HTTP requests (e.g., in smoke tests)

# === Task Queuing / Celery ===
celery[redis]~=5.3.6 # Distributed Task Queue with Redis support
redis~=5.0.0         # Python client for Redis

python-multipart
